requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
xgboost
catboost
joblib
numpy
pandas
python-dotenv
torch
torchvision
torchaudi
Jinja2





